pyproject.toml
[project]
name = "rag-eval-masterclass"
version = "0.1.0"
description = "RAG Evaluation Masterclass — Learn to evaluate Retrieval-Augmented Generation systems"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-multipart",
    "llama-index-core",
    "llama-index-embeddings-huggingface",
    "llama-index-llms-openai",
    "python-dotenv",
    "pandas",
    "tqdm",
    "requests",
    "openai",
]
